LPD-56150 Abstract to new file
