+++ modules/apps/export-import/export-import-web/src/main/resources/META-INF/resources/js/components/screens/DetailViewContentRow.tsx
@@ -1,0 +1,20 @@
+import React from 'react';
+import ClayLayout from '@clayui/layout';
+
+interface DetailViewContentRowProps {
+    title?: string;
+    body: React.ReactNode;
+}
+
+export function DetailViewContentRow({title, body}: DetailViewContentRowProps): JSX.Element {
+    return (
+    <div className='sheet-text'>
+        {title && <ClayLayout.ContentCol className='text-body'>
+            <strong>{title}</strong> 
+        </ClayLayout.ContentCol>}
+        <ClayLayout.ContentCol>
+            {body}
+        </ClayLayout.ContentCol>
+    </div>
+    );
+}
